+++ .changeset/tender-mirrors-tap.md
@@ -1,0 +1,5 @@
+---
+'legend-transac': patch
+---
+
+Se añade micro y comandos para "notificaciones" y "liptv"
